+++ changelog.md
@@ -1,15 +1,24 @@
 # v1.0.6
+
 - Added Argon status check
+- Added custom server status checks (You can add your own url to check if is up or down)
 
 # v1.0.5
+
 - Removed Node ID since it's not needed
 - Fixed the mod settings descriptions
 - Attempted to fix the ZOrder bug
+
 # v1.0.4
+
 - Fixed a critical bug where close menu doesn't exist on mobile players, causing the game to crash on launch :panik:
+
 # v1.0.3
+
 - Changed the Boomlings status to check the most liked levels as an endpoint to determine if the server is up or down
+
 # v1.0.2
+
 - *i actually did something this time*
 - Fixed the icon color logic, should now be:
   - <cg>Green if all services are up</c>
@@ -20,7 +29,11 @@
 - Added option to toggle notifications when a service goes down
 - Added option to customize the URL used for the Internet check
 - Internal Internet check is now only visible by mobile users
+
 # v1.0.1
+
 - i did absolutely nothing
+
 # v1.0.0
+
 - [get wifi anywhere you go HOLD UP vine](https://www.youtube.com/watch?v=9p0pdiTOlzw)